fix: review batch — config wiring, bf16 inference, EMA eval, n_gt clamp
Fresh full-codebase review (multi-team shared module). Fixes the genuinely
needed findings; deferred items (assigner top-k eps, distance match order,
checkpoint-tool fuzzy guards) are intentionally left for separate, validated
changes since they alter training dynamics or live only on opt-in paths.

- configs: loader silently dropped detection_generator.score_thresh, so YAML
  overrides were ignored (pinned to the 0.05 default). Now wired through.
- configs/model: wire task min_distance/max_distance into the detection
  generator's inference distance clamp (single source of truth = task config);
  previously a custom range was honoured only by the loss/parser, not at
  inference/export.
- models/detection_generator: cast raw head outputs to float32 before decode.
  Under the supported mixed_bfloat16 policy the head emits bfloat16, which
  crashed the decode (bf16 x float32) and the float32 fn_output_signature.
  No-op under the default float32 policy.
- tools/continuous_eval: load EMA weights via restore_eval_weights instead of
  raw model/ weights, matching tools/eval.py and the trainer (was reporting a
  non-EMA, noisier curve teams watch for checkpoint selection).
- data_pipeline parsers: clamp stored n_gt to max_num_instances. _pad_labels
  truncates the label tensors, so an image with >max_num_instances objects left
  n_gt larger than the padded width and the loss built a mask wider than the
  tensors → shape crash.
- eval/polygon_metrics: the radial radius is isotropic, so per-axis pixel
  decode is only correct for square inputs. Fixed the docstring (it claimed
  non-square was self-consistent) and now raise on non-square instead of
  silently reporting distorted mIoU.

Tests: added regression tests for the loader wiring and the n_gt clamp.
Full suite 244 passed / 3 skipped.

+++ models/detection_generator.py
@@ -233,8 +233,12 @@ class YoloV8Layer:
 
         for level in ["3", "4", "5"]:
             stride = _LEVEL_STRIDES[level]
-            box_raw = raw_outputs["box"][level]         # [B, H, W, 4*reg_max]
-            cls_raw = raw_outputs["cls"][level]         # [B, H, W, num_classes]
+            # Cast to float32 up front: under a mixed_bfloat16 policy the head
+            # emits bfloat16, which would otherwise clash with the float32
+            # constants/grids below and the float32 fn_output_signature. No-op
+            # under the default float32 policy.
+            box_raw = tf.cast(raw_outputs["box"][level], tf.float32)  # [B, H, W, 4*reg_max]
+            cls_raw = tf.cast(raw_outputs["cls"][level], tf.float32)  # [B, H, W, num_classes]
 
             B = tf.shape(box_raw)[0]
             fH = tf.shape(box_raw)[1]
@@ -263,12 +267,12 @@ class YoloV8Layer:
             all_cls.append(cls_flat)
 
             if has_poly:
-                all_poly_angle.append(tf.reshape(raw_outputs["poly_angle"][level], [B, -1, self.output_poly_size]))
-                all_poly_dist.append( tf.reshape(raw_outputs["poly_dist"][level],  [B, -1, self.output_poly_size]))
-                all_poly_conf.append( tf.reshape(raw_outputs["poly_conf"][level],  [B, -1, self.output_poly_size]))
+                all_poly_angle.append(tf.cast(tf.reshape(raw_outputs["poly_angle"][level], [B, -1, self.output_poly_size]), tf.float32))
+                all_poly_dist.append( tf.cast(tf.reshape(raw_outputs["poly_dist"][level],  [B, -1, self.output_poly_size]), tf.float32))
+                all_poly_conf.append( tf.cast(tf.reshape(raw_outputs["poly_conf"][level],  [B, -1, self.output_poly_size]), tf.float32))
 
             if has_dist:
-                all_dist.append(tf.reshape(raw_outputs["dist"][level][:, :, :, 0], [B, -1]))
+                all_dist.append(tf.cast(tf.reshape(raw_outputs["dist"][level][:, :, :, 0], [B, -1]), tf.float32))
 
         boxes  = tf.concat(all_boxes, axis=1)   # [B, N_anchors, 4]
         scores = tf.concat(all_cls,   axis=1)   # [B, N_anchors, nc]
